feat: make Avatar a standalone component
